+++ src/main/java/miyucomics/hexical/mixin/CircleExecutionStateMixin.java
@@ -4,6 +4,7 @@ import at.petrak.hexcasting.api.casting.circles.BlockEntityAbstractImpetus;
 import at.petrak.hexcasting.api.casting.circles.CircleExecutionState;
 import at.petrak.hexcasting.api.misc.Result;
 import com.llamalad7.mixinextras.sugar.Local;
+import miyucomics.hexical.inits.HexicalBlocks;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -19,8 +20,10 @@ import java.util.ArrayList;
 @Mixin(CircleExecutionState.class)
 public class CircleExecutionStateMixin {
 	@Inject(method = "createNew(Lat/petrak/hexcasting/api/casting/circles/BlockEntityAbstractImpetus;Lnet/minecraft/server/network/ServerPlayerEntity;)Lat/petrak/hexcasting/api/misc/Result;", at = @At(value = "INVOKE", target = "Lat/petrak/hexcasting/api/casting/circles/ICircleComponent;possibleExitDirections(Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/block/BlockState;Lnet/minecraft/world/World;)Ljava/util/EnumSet;", shift = At.Shift.AFTER))
-	private static void beaconsIncreaseAmbit(BlockEntityAbstractImpetus impetus, @Nullable ServerPlayerEntity caster, CallbackInfoReturnable<Result<CircleExecutionState, BlockPos>> cir, @Local BlockState block, @Local ArrayList<BlockPos> goodSeenPositions) {
-		if (block.isOf(Blocks.BEACON))
+	private static void infiniteVerticalAmbit(BlockEntityAbstractImpetus impetus, @Nullable ServerPlayerEntity caster, CallbackInfoReturnable<Result<CircleExecutionState, BlockPos>> cir, @Local BlockState block, @Local ArrayList<BlockPos> goodSeenPositions) {
+		if (block.isOf(HexicalBlocks.WITHERED_SLATE)) {
+			goodSeenPositions.add(new BlockPos(impetus.getPos().getX(), -1024, impetus.getPos().getY()));
 			goodSeenPositions.add(new BlockPos(impetus.getPos().getX(), 1024, impetus.getPos().getY()));
+		}
 	}
 }